feat: folder restructure
